refactor(extension): move the webview pages into a module
The pages were private methods taking a Webview, so nothing could render
them without an extension host. They now take the resolved asset URIs as
plain strings, which lets tests exercise the real markup instead of a
copy that drifts.
